+++ src/main/kotlin/de/andrena/tools/altn8th/actions/openRelatedFile/GoToRelatedFileAction.kt
@@ -16,7 +16,7 @@ import de.andrena.tools.altn8th.adapter.File
 import de.andrena.tools.altn8th.adapter.ProjectFiles
 import de.andrena.tools.altn8th.domain.relatedFiles.deduplicate.strategies.DeduplicateRelationsByTakingFirstOccurrence
 import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.fileExtension.FindRelatedFilesByFileExtensionStrategy
-import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.freeRegex.FindRelatedFilesByFreeRegexStrategy
+import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.filenameRegex.FindRelatedFilesByFilenameRegexStrategy
 import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.postfix.FindRelatedFilesByPostfixStrategy
 import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.prefix.FindRelatedFilesByPrefixStrategy
 import de.andrena.tools.altn8th.domain.relatedFiles.group.strategies.GroupByCategoryStrategy
@@ -34,7 +34,7 @@ class GoToRelatedFileAction : AnAction() {
     private val relatedFilesStrategies = listOf(
         FindRelatedFilesByPrefixStrategy(),
         FindRelatedFilesByPostfixStrategy(),
-        FindRelatedFilesByFreeRegexStrategy(),
+        FindRelatedFilesByFilenameRegexStrategy(),
         FindRelatedFilesByFileExtensionStrategy()
     )