+++ src/main/kotlin/de/andrena/tools/altn8th/domain/relatedFiles/find/strategies/filenameRegex/FindRelatedFilesByFilenameRegexStrategy.kt
@@ -1,23 +1,23 @@
-package de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.freeRegex
+package de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.filenameRegex
 
 import de.andrena.tools.altn8th.domain.File
 import de.andrena.tools.altn8th.domain.relatedFiles.Relation
 import de.andrena.tools.altn8th.domain.relatedFiles.find.strategies.FindRelatedFilesStrategy
 import de.andrena.tools.altn8th.domain.settings.SettingsState
-import de.andrena.tools.altn8th.domain.settings.types.FreeRegexSetting
+import de.andrena.tools.altn8th.domain.settings.types.FilenameRegexSetting
 
-internal class FindRelatedFilesByFreeRegexStrategy : FindRelatedFilesStrategy {
+internal class FindRelatedFilesByFilenameRegexStrategy : FindRelatedFilesStrategy {
     override fun find(origin: File, file: File, settings: SettingsState): Relation? =
         isRelatedBy(settings, origin, file)
-            ?.let { Relation(file, origin, FreeRegexRelationType(it)) }
+            ?.let { Relation(file, origin, FilenameRegexRelationType(it)) }
 
     private fun isRelatedBy(
         settings: SettingsState,
         origin: File,
         file: File
-    ): FreeRegexSetting? {
+    ): FilenameRegexSetting? {
         val regexOptions = if (settings.caseInsensitiveMatching) setOf(RegexOption.IGNORE_CASE) else emptySet()
-        return settings.freeRegexes
+        return settings.filenameRegexes
             .filter { origin.nameWithFileExtension().matches(Regex(it.origin, regexOptions)) }
             .firstOrNull { file.nameWithFileExtension().matches(Regex(it.related, regexOptions)) }
     }