+++ src/main/kotlin/de/andrena/tools/altn8th/settings/ui/components/general/GeneralSettingsUi.kt
@@ -10,14 +10,18 @@ import de.andrena.tools.altn8th.settings.ui.Ui
 import de.andrena.tools.altn8th.settings.ui.multilineContextHelp
 import javax.swing.JPanel
 
-private val TITLE = I18n.lazyMessage("altn8.settings.general")
-private val DESCRIPTION = I18n.lazyMessage("altn8.settings.general.description")
-private val CASE_INSENSITIVE_MATCHING_LABEL = I18n.lazyMessage("altn8.settings.general.caseInsensitiveMatching")
-
-private const val SPACING = 8
 
 internal class GeneralSettingsUi(private val settingsState: SettingsState) : Ui {
-    private val caseInsensitiveMatchingCheckBox = JBCheckBox(CASE_INSENSITIVE_MATCHING_LABEL.get(), settingsState.caseInsensitiveMatching)
+    companion object {
+        private val TITLE = I18n.lazyMessage("altn8.settings.general")
+        private val DESCRIPTION = I18n.lazyMessage("altn8.settings.general.description")
+        private val CASE_INSENSITIVE_MATCHING_LABEL = I18n.lazyMessage("altn8.settings.general.caseInsensitiveMatching")
+
+        private const val SPACING = 8
+    }
+
+    private val caseInsensitiveMatchingCheckBox =
+        JBCheckBox(CASE_INSENSITIVE_MATCHING_LABEL.get(), settingsState.caseInsensitiveMatching)
 
     override val panel: JPanel =
         JPanel(VerticalLayout(SPACING)).apply {
@@ -29,7 +33,8 @@ internal class GeneralSettingsUi(private val settingsState: SettingsState) : Ui 
         panel.border = IdeBorderFactory.createTitledBorder(TITLE.get(), false)
     }
 
-    override fun isModified(): Boolean = caseInsensitiveMatchingCheckBox.isSelected != settingsState.caseInsensitiveMatching
+    override fun isModified(): Boolean =
+        caseInsensitiveMatchingCheckBox.isSelected != settingsState.caseInsensitiveMatching
 
     override fun apply() {
         settingsState.caseInsensitiveMatching = caseInsensitiveMatchingCheckBox.isSelected