+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Navigate to related files using a shortcut, which is (Alt)(8) by default.
 
 ### Changed
 
+* Rename free regex to filename regex to distinguish from filepath regex
+
 ### Fixed
 
 ### Removed